requirements.txt
arrow==1.3.0
cartopy==0.24.1
diffusers==0.32.2
earthkit==0.9.4
einops==0.8.1
gcsfs==2024.10.0
hydra-core==1.3.2
imageio==2.36.0
lightning_utilities==0.12.0
matplotlib==3.7.3
numpy==1.25.2
omegaconf==2.3.0
packaging==24.2
pandas>2.1
Pillow==11.1.0
protobuf==3.20.3
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.15.2
timm==1.0.15
torch==2.6.0
torch_xla==2.6.0
tqdm==4.67.0
xarray==2025.1.2
rasterio==1.4.3
taming-transformers==0.0.1
